PeriodicLogger can be started at a fixed Date
git-svn-id: https://svn.apache.org/repos/asf/commons/sandbox/monitoring/trunk@648190 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/monitoring/listeners/HistorizedRepositoryDecorator.java
@@ -44,7 +44,7 @@ public class HistorizedRepositoryDecorator
 
     private int size;
 
-    public HistorizedRepositoryDecorator( long period, final int size, Observable repository )
+    public HistorizedRepositoryDecorator( int period, final int size, Observable repository )
     {
         super( period, repository );
         this.repository = repository;
